# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AUG22pt-ep1-dissolution2",
    "description": "<b>Swagger Doc EP Noroff Backend API</b>."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "scheme": "bearer",
      "description": "Bearer token"
    }
  },
  "paths": {
...
    "/orders/": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "Admin get all orders, User get her/his orders [token required]",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { orders: \"Orders found\", orders }"
          },
          "403": {
            "description": "statusCode: 403, result: {error: \"Privileges Admin or user rights [Required] !!\"}"
          }
        }
      },
      "put": {
        "tags": [
          "Order"
        ],
        "description": "Admin update/change order status [token required]",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderPut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { result: \"Orders to be updated\", updateOrderList, orderList }"
          },
          "403": {
            "description": "statusCode: 403, result: {error: \"Privileges Admin or user rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode: 500, result: { error: \"Require statusId, Not a integer value, value between 1 - 3\", input: statusId }"
          }
        }
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "description": "Admin can delete a Order by OrderNumber (All tables that are created must have CRUD endpoints)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode\": 200, data: { result: \"Order deleted from database\"}"
          },
          "403": {
            "description": "statusCode\": 403, result: {error: \"Privileges Admin rights [Required] !!\"}"
          },
          "404": {
            "description": "statusCode\": 404 data: { error: \"OrderNumber not found\", input: orderNumber }"
          },
          "500": {
            "description": "statusCode\": 500, result: {error: \"Require orderNumber\", input: orderNumber }"
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "description": "Admin can change a cart to an Order. -Orders are usally Made throue /cart/checkout !! extra !!",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { result: \"Order created\"}"
          },
          "403": {
            "description": "statusCode: 403, result: {error: \"Privileges Admin rights [Required] !!\"}"
          },
          "404": {
            "description": "statusCode: 404, result: {error: \"No product in cart\" }"
          },
          "500": {
            "description": "statusCode: 500, result: { email: \"Email is required, check format.\"}"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrderPut": {
      "type": "object",
      "properties": {
        "orderNumber": {
          "type": "string",
          "example": "uuid-string"
        },
        "statusId": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "orderNumber",
        "statusId"
      ]
    },
    "OrderPost": {
      "type": "object",
      "properties": {
        "userEmail": {
          "type": "string",
          "example": "test@test.no"
        }
      },
      "required": [
        "userEmail"
      ]
    },
    "OrderDelete": {
      "type": "object",
      "properties": {
        "orderNumber": {
          "type": "string",
          "example": "uud-string"
        }
      },
      "required": [
        "orderNumber"
      ]
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}